[S07P12D102-108] FIX : 모임 멤버 파일 정리

+++ BE/CommonProject/src/main/java/com/project/common/service/Group/GroupMemberService.java
@@ -7,19 +7,12 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.project.common.dto.Group.GroupBasicReqDto;
-import com.project.common.dto.Group.GroupDto;
 import com.project.common.dto.Group.GroupJoinReqDto;
-import com.project.common.dto.Group.GroupMemberDto;
 import com.project.common.dto.Group.GroupMemberListDto;
-import com.project.common.dto.Group.GroupSettingDto;
 import com.project.common.entity.Group.GroupEntity;
 import com.project.common.entity.Group.GroupMemberEntity;
-import com.project.common.mapper.GroupMapper;
-import com.project.common.mapper.GroupMemberMapper;
 import com.project.common.repository.Group.GroupMemberRepository;
 import com.project.common.repository.Group.GroupRepository;
-import com.project.common.repository.User.UserRepository;
-import com.project.common.service.UserService;
 
 import lombok.RequiredArgsConstructor;
 
@@ -28,24 +21,21 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class GroupMemberService{
 	private final GroupRepository groupRepository;
-	private final UserRepository userRepository;
 	private final GroupMemberRepository groupMemberRepository;
-	private final UserService userService;
 	private final GroupService groupService;
 
-	
-	//그륩 찾기(그륩 번호로)
-	public List<GroupMemberEntity> findGroup(long groupSeq) {
-		List<GroupMemberEntity> findGroup = new ArrayList<>();
+	//멤버 찾기(그륩 번호로)
+	public List<GroupMemberEntity> findMember(long groupSeq) {
+		List<GroupMemberEntity> findMember = new ArrayList<>();
 		for(GroupMemberEntity entity: groupMemberRepository.findAll()) {
 			if(entity.getGroup().getGroupSeq()==groupSeq) {
-				findGroup.add(entity);
+				findMember.add(entity);
 			}
 		}
-		return findGroup;
+		return findMember;
 	}
 	
-	//가입자 목록 목록 
+	//멤버 목록
 	public List<GroupMemberListDto> getMemberList(long groupSeq){
 		List<GroupMemberListDto> list = new ArrayList<>();
 		for(GroupMemberEntity entity : groupMemberRepository.findAll()) {
@@ -56,7 +46,7 @@ public class GroupMemberService{
 		return list;
 	}
 	
-	//그륩 참가
+	//모임 참가
 	@Transactional
 	public void joinGroup(long groupSeq, GroupJoinReqDto groupJoinRequestDto) {
 		GroupEntity group = groupService.findGroup(groupSeq);
@@ -64,20 +54,22 @@ public class GroupMemberService{
 		groupRepository.save(group);
 	}
 	
-	//그륩 탈퇴
+	//모임 탈퇴
 	@Transactional
 	public void leaveGroup( GroupBasicReqDto groupLeave) {
-		List<GroupMemberEntity> group = findGroup(groupLeave.getGroupSeq());
+		List<GroupMemberEntity> member = findMember(groupLeave.getGroupSeq());
 		GroupEntity groupE = groupService.findGroup(groupLeave.getGroupSeq());
-		for(GroupMemberEntity entity : group) {
+		for(GroupMemberEntity entity : member) {
+			if(entity.getUserSeq()==groupLeave.getUserSeq()) {
 				groupMemberRepository.deleteByUserSeq(groupLeave.getUserSeq());
 				groupE.removeGroupMember(groupLeave.getUserSeq());
+			}
 		}
 	}
 	
-	//모임 수락
+	//모임 가입 승인
 	public void approveMember(Long groupSeq,GroupBasicReqDto groupBasicReqDto) {
-		for(GroupMemberEntity entity: findGroup(groupSeq)) {
+		for(GroupMemberEntity entity: findMember(groupSeq)) {
 			if(entity.getUserSeq()==groupBasicReqDto.getUserSeq()) {
 				entity.setMemberStatus(1);
 				groupMemberRepository.save(entity);
